Report on Edge Detection in Image Processing
The main purpose of edge detection is to simplify the image data to minimize the amount of data to be processed. Image processing works by detecting discontinuities in brightness and is used for image segmentation and data extraction in image processing areas.
Image processing has applications in different fields such as medicines, astronomy, product quality, industrial applications, and plays an important role in the segmentation and object identification process. The software used is Matlab 16.0.