SMART STREET LIGHT SYSTEM USING IR SENSOR
The system is built in such a way that the lights turn ON when necessary and turn
OFF when not required. Currently, huge amounts of electrical energy are being
consumed by the usage of street lights alone. So there is huge wastage of energy on
doing so.
As far as now, LED is considered a promising solution to modern street lighting systems due to its behavior and advantages. like power consumption is less compared to others and life time more compared to others.